Force TLS 1.2 and fix VLAN caching
Fixed the issue with session resumption by adding the Filter_Id to the EAP cache.  Removed the unneeded cache calls from the inner-tunnel and added in cache_ldap_user_dn to make sure the group attributes are cached and recalled.  Changed cache_ldap_user_dn to cache only the LDAP-Group and the Filter-Id